Added .pom and the start of the engine
Added CompressedColor. A class that compresses colors in the RGB
colorspace, and which provides a hash code which gives unique values
for all RGB values.

Added ImageProcessor. This class should return which colors are in an
image, and their relative frequency. Should have a threshold function,
and allow only returning colors with a relative frequency over that
threshold.

Added Test CompressedColorTest, which tests the CompressedColor class.